% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grubbsTest.R
\name{grubbsTest}
\alias{grubbsTest}
\title{Grubbs Test for one outlier in a normal sample}
\usage{
grubbsTest(x, alternative = c("max", "min", "two.sided"), alpha = 0.05)
}
\arguments{
\item{x}{vector of normally distributed data values}

\item{alternative}{character value  specifying the alternative hypothesis.
It must be one of "max" (default), "min" or "two.sided".}

\item{alpha}{significance level for the rejection, acceptable Type 1 error.}
}
\value{
object of class \code{htest} (hypothesis test), i.e. list with test-specific estimates, critical values and messages.
}
\description{
The original paper states to reject the absolute maximum deviate G
at a significance level of alpha/n for one-sided tests and alpha/(2n) for
two sided tests.
}
\examples{
# example code
iso <- c(199.31, 199.53, 200.19, 200.82, 201.92, 201.95, 202.18, 245.57)
grubbsTest(iso)
}
\references{
Grubbs, F. (1950).
\emph{Sample Criteria for Testing Outlying Observations.}
\emph{Annals of Mathematical Statistics,} 21(1), 27–58.
\url{https://doi.org/10.1214/aoms/1177729885}
}
