% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGrubbs.R
\name{qGrubbs}
\alias{qGrubbs}
\alias{pGrubbs}
\title{The Grubbs G Distribution}
\usage{
qGrubbs(p, n, two.sided, lower.tail = FALSE)

pGrubbs(x, n, two.sided, lower.tail = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{n}{sample size.}

\item{two.sided}{logical; if TRUE, an outlier can be either a minimum or a maximum.}

\item{lower.tail}{logical; if FALSE (default), probabilities are
\eqn{P(G > g)}, otherwise \eqn{P(G \le g)}.}

\item{x}{vector of quantiles.}
}
\description{
Quantile function and distribution function for Grubbs G distribution.
}
\references{
Grubbs, F. (1950).
\emph{Sample Criteria for Testing Outlying Observations.}
\emph{Annals of Mathematical Statistics,} 21(1), 27–58.
\url{https://doi.org/10.1214/aoms/1177729885}
}
