% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grubbsTest.R
\name{grubbsTest}
\alias{grubbsTest}
\title{Grubbs Test for one outlier in a normal sample}
\usage{
grubbsTest(x, alternative = c("max", "min", "two.sided"), conf.level = 0.95)
}
\description{
The original paper states to reject the absolute maximum deviate G
at a significance level of alpha/n for one-sided tests and alpha/(2n) for
two sided tests.
}
\examples{
# example code
x <- c(199.31, 199.53, 200.19, 200.82, 201.92, 201.95, 202.18, 245.57)
grubbsTest(x)
}
