% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGrubbs.R
\name{qGrubbs}
\alias{qGrubbs}
\alias{pGrubbs}
\title{The Grubbs G Distribution}
\usage{
qGrubbs(p, n, two.sided, lower.tail = FALSE)

pGrubbs(x, n, two.sided, lower.tail = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{n}{sample size.}

\item{two.sided}{logical; if TRUE, an outlier can be either a minimum or a maximum.}

\item{lower.tail}{logical; if FALSE (default), probabilities are
\eqn{P(G > g)}, otherwise \eqn{P(G \le g)}.}

\item{x}{vector of quantiles.}
}
\description{
Quantile function and distribution function for the extreme absolute deviate G of a
normal sample defined as \eqn{ G = \max (G_1, G_n),}
where \eqn{G_i = \frac{|x_{(i)} - \bar{x}|}{s}} for an ordered sample of
\eqn{x_{i}} with \eqn{x_{(1)} \le x_{(2)} \le ... \le x_{(n)}}.
\eqn{\bar{x}} and \eqn{s^2} are sample mean and variance defined as
\deqn{\bar{x} = \frac{1}{n} \sum_{i=1}^n x_i text{ and } s^2 = \frac{1}{n-1}\sum_{i=1}^n (x_i - \bar{x})^2.}
}
\details{
The quantile and distribution function is derived from
rejecting
\deqn{}
}
